requirements.txt
Django==5.0.1
chromadb==0.4.22
openai==1.11.0
djangorestframework==3.14.0
langchain==0.1.5
langchain-openai==0.0.5
pypdf==4.0.1
docx2txt==0.8
chardet==5.2.0
pikepdf
psycopg2-binary
gunicorn==21.2.0
ipython
celery==5.3.6
redis==5.0.3
debugpy
watchdog
pinecone-client
langchain-pinecone 